// src/lib/api/schema.d.ts
export interface components {
    schemas: {
        CreateMovieRequest: {
            title: string;
            genre?: string[] | null;
            /** Format: int32 */
            year?: number;
            /** Format: double */
            rating?: number;
            /** Format: int32 */
            popularity?: number;
            description?: string | null;
        };
        MetaResponse: {
            /** Format: int32 */
            limit?: number;
        };
        MovieResponse: {
            id?: string | null;
            title?: string | null;
            genre?: string[] | null;
            /** Format: double */
            rating?: number;
            /** Format: int32 */
            year?: number;
            /** Format: int32 */
            popularity?: number;
            description?: string | null;
        };
        MovieResponseApiResponse: {
            data?: components["schemas"]["MovieResponse"];
            meta?: components["schemas"]["MetaResponse"];
        };
        MovieResponseIEnumerableApiResponse: {
            data?: components["schemas"]["MovieResponse"][] | null;
            meta?: components["schemas"]["MetaResponse"];
        };
        ProblemDetails: {
            type?: string | null;
            title?: string | null;
            /** Format: int32 */
            status?: number | null;
            detail?: string | null;
            instance?: string | null;
        } & {
            [key: string]: unknown;
        };
        UpdateMovieRequest: {
            title?: string | null;
            genre?: string[] | null;
            /** Format: int32 */
            year?: number | null;
            /** Format: double */
            rating?: number | null;
            /** Format: int32 */
            popularity?: number | null;
            description?: string | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}